增加活动管理主流程代码

+++ src/main/java/com/haisheng/framework/testng/bigScreen/jiaochen/wm/sense/pc/activity/ManageRegisterApprovalScene.java
@@ -13,7 +13,7 @@ import java.util.List;
 public class ManageRegisterApprovalScene extends BaseScene {
     private final Long activityId;
     private final Integer status;
-    private final List<Integer> ids;
+    private final List<Long> ids;
 
     @Override
     public JSONObject getJSONObject() {
